requirements.txt
coverage==7.2.3
dvc==2.54.0
dvc_s3==2.22.0
numpy==1.24.2
pandas==2.0.0
scikit-learn==1.2.2
nltk==3.8.1
greenplum-python==1.0.1
hvac==1.1.0
kafka-python==2.0.2